@@ -1,14 +1,7 @@
 plugins {
-    id 'java'
-    id 'org.springframework.boot' version '2.6.3'
+    id 'org.springframework.boot' version '3.0.0'
     id 'io.spring.dependency-management' version '1.1.0'
-}
-
-group = 'com.example'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    sourceCompatibility = '11'
+    id 'java'
 }
 
 repositories {
@@ -16,12 +9,12 @@ repositories {
 }
 
 dependencies {
-    runtimeOnly('org.apache.tomcat.embed:tomcat-embed-jasper')
-    runtimeOnly('jakarta.servlet.jsp.jstl:jakarta.servlet.jsp.jstl-api')
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    runtimeOnly 'org.apache.tomcat.embed:tomcat-embed-jasper'
+    runtimeOnly 'jakarta.servlet.jsp.jstl:jakarta.servlet.jsp.jstl-api'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
 
-tasks.named('test') {
-    useJUnitPlatform()
-}
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
+compileJava.options.encoding = 'UTF-8'
